Retry Cloudformation StackSet Create when the service is busy
Stack Sets only let a single stack per account be modified at the same
time. When an operation is in progress, an
`OperationInProgressException` is thrown that's similar to rate limiting
errors. This adds that exception type to the built-in Retryer in the AWS
SDK so it gets auto-retried like rate limits do